requirements.txt
# AI Employee - Silver Tier Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Browser automation (LinkedIn Watcher)
playwright>=1.40.0

# File system watching
watchdog>=3.0.0

# MCP Server framework (optional)
# mcp>=0.1.0
